Split CreamPiedVisualizer into Component/System pair
